separating property pane out of widgets editor to avoid triggering unwanted events on widget canvas
